requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-google-genai
langchain-chroma
chromadb
pypdf